Restructure code, distinguish cone legs from arbitrary base edges
